move skeleton to corresponding component
